requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
google-generativeai>=0.3.0
google-genai>=0.2.0
google-auth>=2.23.0
markdown>=3.5.0
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
openai>=1.0.0
anthropic>=0.18.0
yt-dlp>=2024.0.0
beautifulsoup4>=4.12.0
Pygments>=2.15.0
supabase>=2.0.0
lxml>=4.9.0
urllib3>=2.0.0
